fixup bad variable size
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@39863 72102866-910b-0410-8b05-ffd578937521

+++ Tests/base/NSInvocationOperation/basic.m
@@ -13,7 +13,7 @@ int main()
   NSInvocation          *inv1;
   NSInvocation          *inv2;
   NSValue               *val;
-  int                   length;
+  NSInteger             length;
   NSString              *hello = @"hello";
   NSString              *uppercaseHello;
   NSOperationQueue      *queue = [NSOperationQueue new];
